% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_ale.R
\name{mcmc_ale}
\alias{mcmc_ale}
\title{Accumulated Local Effects for a Bayesian Posterior}
\usage{
mcmc_ale(
  X,
  model,
  pred_fun,
  J,
  K = 40,
  CrI = c(0.025, 0.975),
  f_true = NULL,
  include_main_effects = TRUE,
  center = "avg_pred"
)
}
\arguments{
\item{X}{The data frame of predictor variables to which the model was fit.}

\item{model}{The fitted Bayesian model object from which predictions can be made.}

\item{pred_fun}{A user-supplied function that will be used to predict the response
for \code{model} for some specified input. \code{pred_fun} has two arguments. The first argument
is named \code{model} and the second argument is named \code{newdata}. The output of \code{pred_fun}
when applied to \code{model} and \code{X} must be a numeric matrix \code{nrow(X)} rows and number of
columns equal to the number of posterior samples drawn (this is the main difference
between this function and \code{ALEPlot} from the \href{https://cran.r-project.org/web//packages//ALEPlot/ALEPlot.pdf}{\code{ALEPlot}} package).}

\item{J}{same as in \href{https://cran.r-project.org/web//packages//ALEPlot/ALEPlot.pdf}{\code{ALEPlot}}.}

\item{K}{same as in \href{https://cran.r-project.org/web//packages//ALEPlot/ALEPlot.pdf}{\code{ALEPlot}}.}

\item{CrI}{a numeric vector of length two indicating the pointwise quantiles
to return in addition to the point estimate.}

\item{f_true}{an optional function which maps \code{X} to its true predicted value
(useful for comparing results in a simulation study).}

\item{include_main_effects}{logical determining whether the centered main effect
ALEs should be added back to the second-order ALE (has no effect when \code{dim(J) = 1})}

\item{center}{centring method to be used. Must be one of 'avg_pred' or 'median'.
By default, ALE is centered with respect to the predictor value(s) that corresponds
to the average (mean) ALE prediction. Setting to 'median' will have the effect
of centering the ALE about the ALE prediction for the median predictor value(s).}
}
\description{
This function provides pointwise first- and second-order ALE with uncertainty
intervals using the posterior samples of predictions from a Bayesian model. The
function currently only supports computing ALEs for numeric predictors.
}
\details{
See the Apley (2020) reference below for a background on ALE.
}
\references{
Daniel W. Apley, Jingyu Zhu, Visualizing the Effects of Predictor Variables in
Black Box Supervised Learning Models, Journal of the Royal Statistical Society
Series B: Statistical Methodology, Volume 82, Issue 4, September 2020, Pages
1059–1086, \url{https://doi.org/10.1111/rssb.12377}
}
